Initial commit
This is the initial commit of this project. This commit includes the
basic prototype of the typing test. The application will generate a
random text with set length and language that user has to type out with
as little mistakes as possible and as fast as they can when they finish
typing out the text the test will reset and you can start again. If you
want to restart the test just press tab.