Timer
Timer mode has been added it can be enabled/disabled in menu under Timer
option. When timer is enabled the test will end when the countdown
reaches 0 seconds. The timer will be visible right after first input of
the test and is placed on the top of the left hand side of test text.